Fixed signed/unsigned comparisions Enabled -Wsign-compare, so they are picked up before problems arise Removed -fconserve-space from flags, due to issues with gcc 3.3

+++ src/include/settings.h
@@ -78,7 +78,7 @@ struct _settings_ {
     int		RevealMap;		/// Reveal map
 };
 
-#define SettingsPresetMapDefault	(~0U)	/// Special: Use pud/cm supplied
+#define SettingsPresetMapDefault	(int)(~0U)	/// Special: Use pud/cm supplied
 
 /*
 **	Resource-Preset factor
